+++ driver/lib/driver.ts
@@ -6,7 +6,8 @@ import { IsolateSocket } from './sessions/isolate_socket';
 import { log as logger } from './logger';
 
 import { DRIVER_NAME as IOS_DEVICE_NAME } from './sessions/ios';
-import { executeElementCommand } from './sessions/observatory';
+import { executeElementCommand, executeGetVMCommand,
+  executeGetIsolateCommand } from './sessions/observatory';
 import { createSession, deleteSession } from './sessions/session';
 
 import { driverShouldDoProxyCmd, FLUTTER_CONTEXT_NAME,
@@ -53,6 +54,8 @@ class FlutterDriver extends BaseDriver {
 
   // session
   public executeElementCommand = executeElementCommand;
+  public executeGetVMCommand = executeGetVMCommand;
+  public executeGetIsolateCommand = executeGetIsolateCommand;
   public execute = execute;
   public executeAsync = execute;
 
@@ -155,7 +158,7 @@ class FlutterDriver extends BaseDriver {
     // In WebView context, all request should got to each driver
     // so that they can handle http request properly.
     // On iOS, WebVie context is handled by XCUITest driver while Android is by chromedriver.
-    // It measn XCUITest driver should keep the XCUITest driver as a proxy,
+    // It means XCUITest driver should keep the XCUITest driver as a proxy,
     // while UIAutomator2 driver should proxy to chromedriver instead of UIA2 proxy.
     return this.proxyWebViewActive && this.proxydriverName !== IOS_DEVICE_NAME;
   }