+++ driver/lib/sessions/observatory.ts
@@ -1,9 +1,11 @@
 import { URL } from 'url';
-
+import _ from 'lodash';
 import { deserialize } from '../../../finder/nodejs/lib/deserializer';
 import { FlutterDriver } from '../driver';
 import { log } from '../logger';
 import { IsolateSocket } from './isolate_socket';
+
+const truncateLength = 500;
 
 // SOCKETS
 export const connectSocket = async (
@@ -97,6 +99,7 @@ export const connectSocket = async (
           socket.close();
           return;
         }
+        // e.g. 'isolates/2978358234363215', '2978358234363215'
         socket.isolateId = mainIsolateData.id;
         // @todo check extension and do health check
         const isolate = await socket.call(`getIsolate`, {
@@ -135,6 +138,30 @@ export const connectSocket = async (
   }
   retryCount = 0;
   return connectedSocket;
+};
+
+export const executeGetIsolateCommand = async function(
+  this: FlutterDriver,
+  isolateId: string|number
+) {
+  log.debug(`>>> getIsolate`);
+  const isolate = await this.socket!.call(`getIsolate`, { isolateId: `${isolateId}` });
+  log.debug(`<<< ${_.truncate(JSON.stringify(isolate), {
+    'length': truncateLength, 'omission': `...` })}`);
+  return isolate;
+};
+
+export const executeGetVMCommand = async function(this: FlutterDriver) {
+  log.debug(`>>> getVM`);
+  const vm = await this.socket!.call(`getVM`) as {
+    isolates: [{
+      name: string,
+      id: number,
+    }],
+  };
+  log.debug(`<<< ${_.truncate(JSON.stringify(vm), {
+    'length': truncateLength, 'omission': `...` })}`);
+  return vm;
 };
 
 export const executeElementCommand = async function(